Add dom updates file
